Abstract parseText to a util function
